Fixes following an 0x Alpha AI-driven security review, focused on security, reliability, and correctness. (#15)
The changes introduce stricter validation and fail-closed behavior for truncated, corrupt, or hostile files across JPEG, PNG, GIF, WebP, BMFF, JXL, CR3, TIFF, and XMP. This prevents silent truncation, unbounded allocations, incorrect-size writes, and out-of-memory conditions caused by malicious inputs.

Resource and cancellation handling has also been hardened through bounded RIFF buffers, streaming, Long positions for files larger than 2 GiB, capped zlib output, proper resource closing, and try/finally cleanup.

Platform behavior is now more consistent across JVM, Android, JS, WASM, Apple, and POSIX, including zlib, Latin-1 handling, time-zone behavior, and NSData conversion.

Several correctness edge cases have been addressed, including date and time sub-seconds, ISO 8601 handling, GPS rational and DDM values, exposure and ISO metadata, finite rational validation, and orientation handling.

Additional edge-case unit tests and goldens have been added, with improved coverage for truncated chunk rejection. The changes also include dead-code removal, example cleanup, license attribution updates, and cleanup of Detekt, Kover, and CI cache issues.

+++ src/commonMain/kotlin/de/stefan_oltmann/kim/model/MetadataSummary.kt
@@ -72,7 +72,7 @@ public data class MetadataSummary(
         get() = if (widthPx == null || heightPx == null)
             0
         else
-            (widthPx * heightPx).div(KimValueFormatter.MEGA_PIXEL_COUNT)
+            (widthPx.toLong() * heightPx).div(KimValueFormatter.MEGA_PIXEL_COUNT.toLong()).toInt()
 
     val locationDisplay: String?
         get() = locationShown?.displayString
